requirements.txt
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1

